Solve problem manual way without f"{v:02d}"
Wanted to see how it can be solved without using the f-string format
specification. It was more trickier because fixing a couple of the test
cases made them work but it made the other ones not work anymore.

So had to figure out a more unified approach to get them all to be
solved.

Slowly figuring out a method of debugging that works for me.